Minimize warnings (#159)
* add clang & gcc warnings & error cmake toolchains

* minimize warnings

* update tests

* make nmtools_cassert macro overrideable to use exception

* fix clang ci run tests

* avoid nmtools_cassert macro redefinition

* fix condition's parenthesis on assert

* fix for werror

* fix for gcc specific Werror 🤦

* fix Werror for bench

* quick fix for constexpr Werror unused value

+++ include/nmtools/utils/to_string.hpp
@@ -132,7 +132,8 @@ namespace nmtools::utils
                 // only add open bracket up to axis n
                 // that is equal to shape[n]-1, starting from last axis
                 for (int ii=(int)size(idx)-1; ii>=0; ii--) {
-                    if (at(idx,ii)==(at(s,ii)-1)) {
+                    // for simplicity just use int
+                    if ((int)at(idx,ii)==(int)(at(s,ii)-1)) {
                         str += "]";
                         // also count how much newline to be printed
                         print_comma++;
